Should probably title it unreleased
So for anyone reading that's curious, why was there an initial category called "political"? Well, I was considering releasing a few political articles on my website, including "peace.html" which was a satire on conservatism I wrote for AP Lang, I figured that if I was going to be potentially using this as a professional page, it would be smart to keep it out for now. It's not because I'm afraid to share my political stances, but because it would likely be inappropriate to potential employers. Perhaps I could release these later, but not for now.